backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1"
description = ""
authors = ["ssyucfa <payk1230@bk.ru>", "vrppaul <pavel@paran.in>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
docxtpl = "^0.12.0"
psycopg2 = "^2.9.1"
python-dotenv = "^0.19.0"
gunicorn = "^20.1.0"
coverage = "^6.0.1"
dj-rest-auth = "^2.1.11"
djangorestframework-simplejwt = "^5.0.0"
django-allauth = "^0.45.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-django = "^4.4.0"
pytest-factoryboy = "^2.1.0"
pytest-cov = "^2.12.1"
pytest-pythonpath = "^0.7.3"
pre-commit = "^2.15.0"
mypy = "^0.910"
flake8 = "^3.9.2"
isort = "^5.9.3"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
